==> 01-scrape.ipynb <==
## Scrape judges' education

import pandas as pd

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.support.ui import Select

from webdriver_manager.chrome import ChromeDriverManager

import requests
from bs4 import BeautifulSoup

import re

driver = webdriver.Chrome(ChromeDriverManager().install())

## DC Court

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-district-columbia-circuit-justices-and-judges")
doc = BeautifulSoup(driver.page_source)

# get links -- list
list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    # this goes through each link in list of links
    myhref = "https://www.fjc.gov"+entry['link']
    # get link
    driver.get(myhref)
    # get doc
    doc = BeautifulSoup(driver.page_source)
    # create empty dictionary
    each_judge = {}
    # for each row, find the name using beautiful soup
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    # split the doc into lines and then find the text between Education: and Professional
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

district_df = pd.DataFrame(the_table)
district_df.head()

len(district_df)

## First Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-first-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit1_df = pd.DataFrame(the_table)
len(circuit1_df)

## Second Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-second-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit2_df = pd.DataFrame(the_table)
len(circuit2_df)

## Third Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-third-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit3_df = pd.DataFrame(the_table)
len(circuit3_df)

## Fourth Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-fourth-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit4_df = pd.DataFrame(the_table)
len(circuit4_df)

## Fifth Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-fifth-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit5_df = pd.DataFrame(the_table)
len(circuit5_df)

## Sixth Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-sixth-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit6_df = pd.DataFrame(the_table)
len(circuit6_df)

## Seventh Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-seventh-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit7_df = pd.DataFrame(the_table)
len(circuit7_df)

## Eighth Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-eighth-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit8_df = pd.DataFrame(the_table)
len(circuit8_df)

## Ninth Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-ninth-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit9_df = pd.DataFrame(the_table)
len(circuit9_df)

## Tenth Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-tenth-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit10_df = pd.DataFrame(the_table)
len(circuit10_df)

## Eleventh Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-eleventh-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

circuit11_df = pd.DataFrame(the_table)
len(circuit11_df)

## Federal Circuit

driver.get("https://www.fjc.gov/history/courts/u.s.-court-appeals-federal-circuit-judges")
doc = BeautifulSoup(driver.page_source)

list_o_links = []
for judge in doc.find_all('a', class_='colorbox-node colorbox-node-gallery init-colorbox-node-processed-processed'):
    judge_dict = {}
    judge_dict['name'] = judge.text
    judge_dict['link'] = judge['href']
    list_o_links.append(judge_dict)

len(list_o_links)

the_table = []
for entry in list_o_links:
    myhref = "https://www.fjc.gov"+entry['link']
    driver.get(myhref)
    doc = BeautifulSoup(driver.page_source)
    each_judge = {}
    for row in doc:
        each_judge['name'] = doc.find_all('h1')[1].text
    
    doc_lines = doc.get_text().splitlines()
    for line in doc_lines:
        if re.search(r"(?:Education[:])(.*)(?:Professional)", line):
            education_line = line
            education_text = re.findall(r"(?:Education[:] )(.*)(?:Professional)",education_line,re.M)
            education_text = education_text[0]
            each_judge['school'] = education_text
    
    the_table.append(each_judge)

len(the_table)

federal_df = pd.DataFrame(the_table)
len(federal_df)

## Combining dataframes

# Create column in each dataframe with the name of the circuit
circuit1_df['circuit'] = 'First'
circuit2_df['circuit'] = 'Second'
circuit3_df['circuit'] = 'Third'
circuit4_df['circuit'] = 'Fourth'
circuit5_df['circuit'] = 'Fifth'
circuit6_df['circuit'] = 'Sixth'
circuit7_df['circuit'] = 'Seventh'
circuit8_df['circuit'] = 'Eighth'
circuit9_df['circuit'] = 'Ninth'
circuit10_df['circuit'] = 'Tenth'
circuit11_df['circuit'] = 'Eleventh'
district_df['circuit'] = 'District of Columbia'
federal_df['circuit'] = 'Federal'

df_list = [circuit1_df, circuit2_df, circuit3_df, circuit4_df, circuit5_df, circuit6_df, circuit7_df, circuit8_df,
         circuit9_df, circuit10_df, circuit11_df, district_df, federal_df]

final_df = pd.concat(df_list)
len(final_df)

final_df.head()

final_df.to_csv('01-scrape.csv', index=False)
