==> 02-split.ipynb <==
## Split education column from the scrape

import pandas as pd
import numpy as np
import re

## Split name and school columns

df_raw = pd.read_csv('01-scrape.csv')
df_raw.head()

# testing splits with string
string_test = "St. John’s College, Annapolis, Maryland, A.B., 1913Harvard Law School, LL.B., 1916St. John’s College, Annapolis, Maryland, A.M., 1917"

print(re.split('(\d+)', string_test))

# split school column 
education_df = df_raw["school"].str.split(r'(\d+)', expand=True)
education_df.head()

# Replace '' with None
education_df = education_df.replace('', None)

# Replace all Nones with NaNs
education_df = education_df.fillna(value=np.nan)

# drop columns that only have Nones 
# only column 10, tested using:
# test_df[10].unique()
# columns 8 & 9 each only have one entry

education_df = education_df.dropna(axis=1, how='all')
education_df.columns

# rename columns
column_indices = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
new_names = ['school1', 'year1', 'school2', 'year2', 'school3', 'year3', 'school4', 'year4', 'school5', 'year5']
old_names = education_df.columns[column_indices]
education_df.rename(columns=dict(zip(old_names, new_names)), inplace=True)

# merge dataframes
i = df_raw.columns.get_loc('school')

df = pd.concat([df_raw.iloc[:, :i], education_df, df_raw.iloc[:, i+1:]], axis=1)
df.head()

# printed this to test that it worked
# df.school1.unique()

# some schools didn't split because they didn't have years, will manually fix these later

# split name columns
name_df = df["name"].str.split(',', expand=True)
name_df.head()

# rename columns
column_indices = [0, 1, 2]
new_names = ['lastname', 'firstname', 'suffix']
old_names = name_df.columns[column_indices]
name_df.rename(columns=dict(zip(old_names, new_names)), inplace=True)

# merge split names df and df with rest of information
df = pd.concat([name_df, df], axis=1)
df.head()

# removing trailing spaces and commas
# test = 'Duquesne University, B.A., '
# line = test.strip().rstrip(',')
# print(line)

df = df.replace(r', $','', regex=True)
df.head()

# in school 1, there are some schools that appeared back to back, like this, because they had no years:
# University of PennsylvaniaBates College, A.B.
# further split the schools that had no years

# testing splits with string
string_test = "University of PennsylvaniaBates College, A.B."

string_new = re.sub(r'([a-z])([A-Z])', r'\1 | \2', string_test)
print(string_new)

print(string_new.split(' | '))

string_test = "University of PennsylvaniaBates College, A.B."

string_new = re.sub(r'([a-z])([A-Z])', r'\1|\2', string_test)
print(string_new)

print(string_new.split('|'))

# Add symbols in between lowercase so as to now break it when I later split
df = df.replace(".*DePau\w.*", "De###Paul University", regex=True)

df.school1 = df.school1.str.replace(r'([a-z])([A-Z])', r'\1|\2', regex=True)

split_df = df.school1.str.split('|', expand=True)
split_df.head(10)

# Remove the ### I added in DePaul
split_df = split_df.replace("###", "", regex=True)

# rename columns
column_indices = [0, 1, 2]
new_names = ['school1_alt', 'school2_alt', 'school3_alt']
old_names = split_df.columns[column_indices]
split_df.rename(columns=dict(zip(old_names, new_names)), inplace=True)

# merge this new df and df with rest of information
df = pd.concat([split_df, df], axis=1)
df.head()

# delete the old school1 column, since these 3 new columns effectively take the place of it
del df['school1']

df.to_csv('02-split.csv', index=False)
