==> 04-format.ipynb <==
## Format data to get a dataframe with the number of times each school appears

import pandas as pd
# pd.set_option('max_rows', None)

df = pd.read_csv('03-clean.csv')
df.head()

long_df = pd.melt(df, id_vars=['name', 'circuit', 'lastname', 'firstname'], value_vars=['school1_alt', 'school2_alt', 'school3_alt', 'school2', 'school3', 'school4', 'school5'], value_name='school')
long_df.head()

# To make a table of people from each school in total, across all circuits

count_all_df = pd.DataFrame(long_df.school.value_counts()).reset_index()
count_all_df.columns = ['school', 'count']

len(count_all_df)

count_by_circuit_df = pd.DataFrame(long_df.groupby('circuit').school.value_counts().reset_index(name='count'))

# rearrange columns to merge with the above dataframe
count_by_circuit_df = count_by_circuit_df[['school', 'count', 'circuit']]

len(count_by_circuit_df)

count_by_circuit_df.to_csv('04-format.csv', index=False)
count_all_df.to_csv('04-format-table.csv', index=False)
