==> 05-address.ipynb <==
## Get addresses for each school

import pandas as pd
import re

# Data downloaded from https://ope.ed.gov/dapip/#/download-data-files
address_df = pd.read_csv('InstitutionCampus.csv')
df = pd.read_csv('04-format.csv')

address_df_filter = address_df[address_df.LocationType == 'Institution']

# Test matches for merge
df['school'].isin(address_df_filter['LocationName']).value_counts()

address_df_filter = address_df_filter.replace("University of Texas at Austin", "University of Texas")
address_df_filter = address_df_filter.replace("University of Nebraska - Kearney", "University of Nebraska")
address_df_filter = address_df_filter.replace("University of South Carolina - Columbia", "University of South Carolina")
address_df_filter = address_df_filter.replace("Florida Agricultural and Mechanical University", "Florida A&M University")
address_df_filter = address_df_filter.replace("Brooklyn College of the City University of New York", "Brooklyn College")
address_df_filter = address_df_filter.replace("Columbia University in the City of New York", "Columbia University")
address_df_filter = address_df_filter.replace("University of Michigan - Ann Arbor", "University of Michigan")
address_df_filter = address_df_filter.replace("University of California - Los Angeles", "University of California, Los Angeles")
address_df_filter = address_df_filter.replace("George Washington University, The", "George Washington University")
address_df_filter = address_df_filter.replace("Indiana University Bloomington", "Indiana University")
address_df_filter = address_df_filter.replace("The University of Alabama", "University of Alabama")
address_df_filter = address_df_filter.replace("SUNY at Buffalo", "State University of New York at Buffalo")
address_df_filter = address_df_filter.replace("University of Maryland - College Park", "University of Maryland")
address_df_filter = address_df_filter.replace("University of North Carolina at Chapel Hill", "University of North Carolina")
address_df_filter = address_df_filter.replace("Louisiana State University and A&M College", "Louisiana State University")
address_df_filter = address_df_filter.replace("University of Minnesota - Twin Cities", "University of Minnesota")
address_df_filter = address_df_filter.replace("University of Wisconsin Colleges", "University of Wisconsin")
address_df_filter = address_df_filter.replace("University of Arkansas at Little Rock", "University of Arkansas")
address_df_filter = address_df_filter.replace("University of Missouri - Columbia", "University of Missouri")
address_df_filter = address_df_filter.replace("Rutgers, The State University of New Jersey", "Rutgers University")
address_df_filter = address_df_filter.replace("University of Illinois Urbana-Champaign", "University of Illinois")
address_df_filter = address_df_filter.replace("Montana State University - Bozeman", "Montana State University")
address_df_filter = address_df_filter.replace("University of Pittsburgh - Main Campus", "University of Pittsburgh")
address_df_filter = address_df_filter.replace("SUNY at Albany", "State University of New York at Albany")
address_df_filter = address_df_filter.replace("The University of Montana", "University of Montana")
address_df_filter = address_df_filter.replace("University of Nevada - Reno", "University of Nevada, Reno")
address_df_filter = address_df_filter.replace("Wittenberg University", "Wittenberg College")
address_df_filter = address_df_filter.replace("Washburn University of Topeka", "Washburn University")
address_df_filter = address_df_filter.replace("University of California Hastings College of Law", "University of California, Hastings College of the Law")
address_df_filter = address_df_filter.replace("Washington University in St Louis", "Washington University in St. Louis")
address_df_filter = address_df_filter.replace("Queens College of the City University of New York", "Queens College, City University of New York")
address_df_filter = address_df_filter.replace("Auburn University Main Campus", "Auburn University")
address_df_filter = address_df_filter.replace("The University of Tennessee - Knoxville", "University of Tennessee")
address_df_filter = address_df_filter.replace("Washington & Jefferson College", "Washington and Jefferson College")
address_df_filter = address_df_filter.replace("California State University - Sacramento", "California State University, Sacramento")
address_df_filter = address_df_filter.replace("SUNY College at Cortland", "State University of New York at Cortland")
address_df_filter = address_df_filter.replace("FLS International St. Peter's University", "St. Peter's College")
address_df_filter = address_df_filter.replace("City College of New York of the City University of New York, The", "City College of New York")
address_df_filter = address_df_filter.replace("Saint Louis University", "St. Louis University")
address_df_filter = address_df_filter.replace("Saint Joseph's University", "St. Joseph's University")
address_df_filter = address_df_filter.replace("University of Massachusetts Amherst", "University of Massachusetts")
address_df_filter = address_df_filter.replace("Saint Josephs College", "St. Joseph's College")
address_df_filter = address_df_filter.replace("North Carolina A & T State University", "North Carolina A&T State University")
address_df_filter = address_df_filter.replace("Southern Illinois University Carbondale", "Southern Illinois University")
address_df_filter = address_df_filter.replace("University of Hawaii at Manoa", "University of Hawaii")
address_df_filter = address_df_filter.replace("Birmingham Southern College", "Birmingham-Southern College")
address_df_filter = address_df_filter.replace("University of Jamestown", "Jamestown College")
address_df_filter = address_df_filter.replace("The University of the South", "University of the South")
address_df_filter = address_df_filter.replace("University of Colorado at Boulder", "University of Colorado")
address_df_filter = address_df_filter.replace("Eastern Kentucky University", "Central University (now Eastern Kentucky University)")

# Replace names that appear as duplicates in "LocationName"
address_df_filter.loc[address_df_filter.DapipId == 140146, "LocationName"] = "Lincoln University, Pennsylvania"
address_df_filter.loc[address_df_filter.DapipId == 125569, "LocationName"] = "Lincoln University, Missouri"
address_df_filter.loc[address_df_filter.DapipId == 126401, "LocationName"] = "Westminster College, Fulton, Missouri"
address_df_filter.loc[address_df_filter.DapipId == 141732, "LocationName"] = "Westminster College, New Wilmington, Pennsylvania"

# Test matches for merge
df['school'].isin(address_df_filter['LocationName']).value_counts()

# Remove duplicates for the merge
address_df_nodups = address_df_filter.drop_duplicates(subset=['LocationName'])

df_merge = pd.merge(df, address_df_nodups, left_on='school', right_on='LocationName', how="left")

# Check length
print(len(df))
print(len(df_merge))

# Look at the places with no match
df_NOT_merge = pd.merge(df, address_df_filter, left_on='school', right_on='LocationName', how="outer",indicator=True)
df_NOT_merge[df_NOT_merge._merge == 'left_only'].school.unique()

# Custom replace remaining addresses
# Leave schools outside of the US blank

# For schools that have closed where addresses couldn't be found, I put their address as the City Hall
# St. Joseph's, Florida, St. Augustine's & Benton College of Law

# Radcliffe College > is now Radcliffe Yard in Harvard
df_merge.loc[df_merge.school == "Radcliffe College", "Address"] = "52 Linnaean St, Cambridge, MA 02138"

df_merge.loc[df_merge.school == "Mitchell Hamline School of Law", "Address"] = "875 Summit Ave, St Paul, MN 55105"

# Crozer Theological Seminary has no address -- is closed now
# Said it's located at W 21st St & Upland Ave so I picked a building near there
df_merge.loc[df_merge.school == "Crozer Theological Seminary", "Address"] = "1100 Upland Ave, Chester, PA 19013"

df_merge.loc[df_merge.school == "Bear Valley Bible Institute", "Address"] = "2707 S. Lamar Street Denver, Colorado 80227"

# St. Augustine's has no address -- is closed now
# https://beniciahistoricalmuseum.org/project/st-augustines-college-c-1878/
df_merge.loc[df_merge.school == "St. Augustine's College", "Address"] = "230E E L St, Benicia, CA 94510"

df_merge.loc[df_merge.school == "Chicago-Kent College of Law", "Address"] = "565 W Adams St, Chicago, IL 60661"

df_merge.loc[df_merge.school == "Judge Advocate General's Legal Center and School", "Address"] = "600 Massie Rd, Charlottesville, VA 22903"

# St. Joseph's College, Florida is also closed
# https://en.wikipedia.org/wiki/Saint_Joseph_College_of_Florida
df_merge.loc[df_merge.school == "St. Joseph's College, Jensen Beach, Florida", "Address"] = "1912 NE Jensen Beach Blvd. Jensen Beach, Florida 34957"

# Douglas University
df_merge.loc[df_merge.school == "St. Joseph’s College, Rensselaer, Indiana", "Address"] = "121 N LaSalle St, Chicago, IL 60602"

# St. Joseph’s College, Rensselaer, Indiana is also closed
# https://www.bloomberg.com/profile/company/1720376Z:US
df_merge.loc[df_merge.school == "Douglas University, Chicago, Illinois", "Address"] = "1498 South College Avenue PO Box 870 Rensselaer, IN 47978"

# Benton College of Law is also closed
# https://lawandhistoryreview.org/article/the-original-failing-law-school-misappropriation-cronyism-and-fisticuffs-at-the-benton-college-of-law/#_ftn4
df_merge.loc[df_merge.school == "Benton College of Law", "Address"] = "1200 Market St, St. Louis, MO 63103"

# Tacoma Law School, according to Wikipedia, is now Seattle University's School of Law https://en.wikipedia.org/wiki/Homer_Bone
df_merge.loc[df_merge.school == "Tacoma Law School", "Address"] = "901 12th Ave, Seattle, WA 98122"

# Central YMCA is also closed
# https://collections.carli.illinois.edu/digital/collection/uic_bru/id/1727/
df_merge.loc[df_merge.school == "Central YMCA College", "Address"] = "211 W Wacker DrChicago, IL 60606"

# Replace addresses that are PO boxes
# University of Washington
df_merge.loc[df_merge.school == "University of Washington", "Address"] = "1400 NE Campus Parkway, Seattle, WA, 98195"

# University of Cincinnati
df_merge.loc[df_merge.school == "University of Cincinnati", "Address"] = "2600 Clifton Ave, Cincinnati, OH 45221"

# Montana State University
df_merge.loc[df_merge.school == "Montana State University", "Address"] = "Culbertson Hall, 100, Bozeman, MT 59717"

# College of William and Mary
df_merge.loc[df_merge.school == "College of William and Mary", "Address"] = "Sadler Center, 200 Stadium Dr, Williamsburg, VA 23185"

# Replace West Virginia State, whose address (Rte 25, Institute, WV 25112-1000) geocodes wrong
df_merge.loc[df_merge.school == "West Virginia State University", "Address"] = "200 Erickson Alumni Ctr, Institute, WV 25112"

# Remove extra columns
df_merge = df_merge[['school','count', 'circuit', 'Address']]
df_merge.columns = df_merge.columns.str.lower()

# Double check nulls
df_merge[df_merge.isnull().any(axis=1)]

df_merge.to_csv('05-address.csv', index=False)
