==> 06-geocode.ipynb <==
## Convert address to lat/lng

import pandas as pd
import requests
from tqdm.auto import tqdm
tqdm.pandas()

df = pd.read_csv('05-address.csv')
df.head()

API_KEY = ''

def geocode(address):

    params = {
        'q': address,
        'apikey': API_KEY
    }

    response = requests.get('https://geocode.search.hereapi.com/v1/geocode', params=params)
    data = response.json()
    if len(data['items']) > 0:
        match = data['items'][0]
        # Create a dictionary of everything important
        result = {
            'lat': match['position']['lat'],
            'lng': match['position']['lng']
        }
        return pd.Series(result)

# use tqdm to get the time
geocoded = df.address.progress_apply(geocode)

geocoded.head()

# Combine dataframes
df_merge = pd.concat([df, geocoded], axis=1)
df_merge.head()

# Check the nulls -- should only be international schools
df_merge[df_merge.isnull().any(axis=1)]

# Replace the NaN lat/lng values with actual NaNs
df_merge['lat'].replace(6.48812, np.nan, inplace=True)
df_merge['lng'].replace(2.6138, np.nan, inplace=True)

# Remove NAs so they don't show up in the final map
df_merge = df_merge.dropna()
# df_merge[df_merge.isnull().any(axis=1)]

# Save file
df_merge.to_csv('06-geocode.csv', index=False)
